Changes to improve security of shiftfs mounts.
These changes are needed in sysbox-runc due to a change in the way sysbox-mgr
will setup shiftfs mounts on host directories/files that are bind-mounted into
containers. Refer to the corresponding change in sysbox-mgr for details.

In sysbox-runc, the changes are mainly to receive from sysbox-mgr the list of
dirs under /var/lib/sysbox/shiftfs/ where the shiftfs mounts have been setup.
sysbox-runc will then bind-mount these into the appropriate directories
in the container.

A benefit of this change is that a couple of shiftfs-mount restrictions that
sysbox-runc was previously imposing no longer apply (e.g., the list of host dirs
in the shiftfs blacklist is no longer required). This gives more flexibility to
users in the host dirs that they can bind-mount into Sysbox containers.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>