Fix symlink existence check in FsEntry.add
The check in FsEntry.add for the SoftlinkFsKind case is supposed to
check if the symlink exists, but it used os.Stat() which will follow
the symlink if it exists and thus check the existence of the target
and not the symlink itself.

Use os.Lstat() instead which will not follow the symlink.

The check caused an error when restarting (docker stop + docker start)
some containers (e.g. Alpine-based) on my Arch Linux system, since the
creation of the symlink to the Linux kernel headers failed.
It doesn't appear to reproduce on Ubuntu nor Fedora, and I'm not 100%
sure of the exact circumstances. Arch may be special because the module
path in Arch is in /lib/modules itself.

Signed-off-by: Joan Bruguera <joanbrugueram@gmail.com>