Fix bug in remount flags.
Per mount(2), a remount operation must keep the original flags
of the mount, except for those that are being changed during
the remount operation.

The remount() function in rootfs_linux.go was not honoring this
requirement, causing remounts to fail in cases where the
filesystem was mounted had non-standard flags. This commit
fixes this.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>