Ensure sysbox mounts in read-only containers are mounted as read only.
In containers that have a read-only rootfs (e.g., docker run --read-only ...),
we must ensure that sysbox's implicit mounts are also read-only. Prior to this
change this was not the case, enabling processes inside the read-only container
to write to portions of the container's filesystem backed by sysbox mounts.

This commit fixes this and adds a test to verify the fix. The fix is not yet
complete, as the sysbox-fs mount over the container's "/proc/sys" continues to
be read-write, even though it was remounted read-only.  We will investigate and
fix in a later commit.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>