Query sysbox-fs for it's fuse mountpoint config.
This change causes sysbox-runc to query sysbox-fs for the location of the fuse
mountpoint.

Prior to this change, sysbox-runc was assuming the sysbox-fs fuse mountpoint was
located under host directory "/var/lib/sysboxfs", but this may not always be
true given that it's configurable in the sysbox-fs command line.

Fixes Sysbox issue #319.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>