Skip shiftfs mounts on bind-mount sources when not required.
sysbox-runc had a bug where it was mounting shiftfs on a bind-mount source
even when it was not required (i.e., when the bind-mount source has
a uid:gid that falls within the host uid:gid range mapped to the sysbox
container via the Linux user-namespace). This commit fixes this.

Fixes sysbox issue #376.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>