Perform uid shifting of Sysbox special dirs on a need basis.
Special dirs are those which must be backed by ext4 inside a sys container (or
more accurately not backed by overlayfs or shiftfs). This include
/var/lib/docker, /var/lib/kubelet, etc.

To do this, Sysbox creates host dirs that it mounts into the sys container's
special dirs. In addition, Sysbox must set the correct ownership on the dirs
such that it matches the sys container's root user.

Prior to this change, how to set that ownership was determined by the ownership
of the container's rootfs. But it's better to make this determination by looking
at the ownership of the special dir itself in the rootfs (if it exists), given
that it's ownership may not match that of the rootfs.

This is in fact necessary in order to fix a problem discovered in Sysbox
issue #254, where running Docker 20.10 inside a sysbox-powered pod caused
the inner Docker to fail with:

failed to register layer: ApplyLayer exit status 1 stdout: stderr: unlinkat /etc/.pwd.lock: operation not permitted

That operation not permitted was caused by an ownership problem on
the inner "/var/lib/docker", due to the fact that inside the pod's image
the rootfs had different ownership than the subdir /var/lib/docker.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>