Fix bug in function that creates dirs under the container's rootfs.
When running a container with CRI-O + Sysbox (i.e., with the user-namespace
enabled), CRI-O typically configures the ownership of the container's
rootfs to match the UID assigned to the container.

However, CRI-O skips this ownership config for subdirs of the container's rootfs
marked as mountpoints in the image file (e.g., via the Dockerfile VOLUME
instruction).

As a result, when the container is being created and in cases where the
container's init process needs to crete a subdir inside the container rootfs,
it's possible that it won't have the permissions to do so (i.e., the container's
init process runs with a non-root UID, yet the image has root ownership in the
parent dir because is designated as a mountpoint). This causes the container to
fail to start.

This change fixes this by having the parent sysbox-runc process (which runs as
root) create the rootfs subdirs, in cases where the container's init process
lacks the permissions to do so.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>