Initial changes to chown the container's rootfs when needed.
The chown is used in Linux hosts that don't support shiftfs.  In this case,
Sysbox will manually "shift" the container's rootfs uid:gids to match the
assigned user-ns uid:gids, using chown.

Since chown can be very expensive when done on top of overlayfs, Sysbox ensures
that the overlayfs is mounted with the "metacopy=on" option (and if not,
remounts it with that option set). This option has a massive impact in reducing
the time it takes to chown the container's rootfs (particularly for heavier
containers).

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>