Add support for using ID-mapped mounts to Sysbox.
This change brings in support for the (long-awaited) ID-mapped mounts feature in
Linux kernels >= 5.12.

This feature is an built-in kernel alternative to the shiftfs module supported
in Ubuntu & Debian distros.

It brings several benefits over shiftfs such as:

* Availability on most Linux distros.
* No need to mark & mount (as required by shiftfs); simply mount.
* Mounting over files or directories (shiftfs only works on directories).
* Mounting over sockets (shiftfs does not work well on sockets).

Some drawbacks are:

* ID-mapped mounts require kernel >= 5.12. Shiftfs works on older kernels (Ubuntu with Kernel >= 5.0).
* ID-mapped-mounts don't work on top of overlayfs yet (shiftfs does).

In this commit, we add the ability for Sysbox to use ID-mapped-mounts when
available, and otherwise fall back to shiftfs or chown as alternative
"filesystem user ID shifting" mechanisms.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>