Wire Sysbox config options to disable shiftfs and ID-mapped-mounts.
These config options are passed to sysbox-runc from sysbox-mgr during
container registration.

They are helpful for testing purposes, as they will allow the Sysbox test suite
to test behavior with/without shiftfs/ID-mapped-mounts on a single host machine
that supports both.

Note that when both shiftfs and ID-mapped-mounts are enabled, Sysbox uses
ID-mapped-mounts. When both are disabled, Sysbox uses chown on the container's
rootfs but not on volume mounts, causing the latter to show up as nobody:nogroup
inside the container.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>