Replace ovfs remount with rootfs cloning.
Rootfs cloning is required when Sysbox needs to chown the rootfs,
and the rootfs is mounted on overlayfs without metacopy=on (as
when using Docker + Sysbox on hosts without shiftfs).

Prior to this change Sysbox was remounting the rootfs with metacopy=on,
but that's incorrect for 2 reasons:

* The remounting requires an unmount+mount, and the unmount appears
  to confuse Docker/containerd.

* Enabling metacopy=on on the rootfs causes the contents of the
  overlayfs upper dir to have file metadata in some cases, rather
  the full file contents. This in turn breaks features such as
  Docker commit that pick up thew diffs from the overlayfs upper dir.

The solution is to have Sysbox clone to rootfs to a new location under
/var/lib/sysbox, mount overlayfs over this new rootfs using the same layers as
the original rootfs, and enable metacopy=on (for fast chowns). Then, Sysbox
bind-mounts the new rootfs merged dir over the original rootfs merged and upper
dirs. This way, rootfs chowns go very fast, and features like Docker commit
work (though the commit layer will in essence pick up an entire new copy
of the container's filesystem, rather than just the diffs, because we are
mounting the merged dir of the new rootfs over the upper dir of
the original rootfs).

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>