Use shiftfs on the rootfs when available.
Prior to this change, if idmapped mounts where supported by the kernel,
sysbox-runc would choose to chown the rootfs (since idmapped mounts
don't work on overlayfs), even if the kernel supported shiftfs.

With this change, sysbox-runc will always choose to use shiftfs on the
container's rootfs when available. This way, it avoids having to chown the
container's rootfs, which is the least desirable option.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>