Add per-container config for Sysbox behavior via container env vars.
This commit adds a mechanism that allows users to configure Sysbox behavior on a
per-container basis, using environment variables passed to the container.

The env vars have the form "SYSBOX_*=VAL". In this commit, two such variables
are supported:

SYSBOX_IGNORE_SYSFS_CHOWN: when set to "TRUE", causes Sysbox to ignore chown of
/sys inside the container (as this otherwise fails due to the container's
user-namespace).

SYSBOX_ALLOW_TRUSTED_XATTR: when set to "TRUE", causes Sysbox to allow setting
of the overlayfs trusted.overlay.opaque attribute within the container. Normally
this is forbidden due to the container's user-namespace.

When these variables are not specified, the behavior is controlled by the
corresponding flags in the sysbox-mgr (e.g., --ignore-sysfs-chown and
--allow-trusted-xattr). That is, the SYSBOX_* env vars override the flags in the
sysbox-mgr.

It is expected that more such env vars will be added in the future, although
it's best to keep them to a minimum because they are Sysbox-specific, thus
reducing compatibility of the container images or container run commands with
other runtimes.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>