Fix shiftfs mounts over files that are themselves bind-mounts.
When a host file is mounted on a sysbox container, Sysbox may need to mount
shiftfs on that host file. But since shiftfs can't be mounted on files (must be
mounted on dirs), Sysbox mounts shiftfs on the parent dir and then bind mounts
the file into the container.

A problem occurs however when that host file is itself a bind-mount from another
file (hereafter "the original file"). In this case, mounting shiftfs on the
parent dir of the host file is not what's needed; instead, Sysbox must mount
shiftfs on the parent dir of the original file.

We've seen this problem occur when using K8s + Sysbox, where K8s mounts secrets
into the sysbox container using files that are themselves bind-mounted from a
different location.

This commit solves this, by having Sysbox check for whether a host files
on which shiftfs will be mounted is itself a bind-mount, and if so
finding the original file associated with that bind-mount and mounting
shiftfs on its parent dir.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>