Use shiftfs as backup for ID-mapped mounts.
Prior to this change, if the host machine's kernel supported both shiftfs and
idmapped-mounts, Sysbox would setup the container's bind mounts with ID-mapped
mounts and not with shiftfs.

However, in some cases ID-mapped mounts don't work (e.g., if the bind mount
source is on overlayfs or btrfs) but shiftfs does. In those cases Sysbox should
use shiftfs for that particular bind-mount, but it wasn't (i.e., it was neither
using idmapped-mounts or shiftfs).

This change fixes this.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>