Reduce syscall trapping when not operating in syscall mode.
Prior to this change, sysbox was trapping the *xattr and chown syscalls even
when not operating in sys container mode. This is not correct and could
negatively affect performance of microservice containers deployed with Sysbox.

This change correct this. Note that some syscalls (e.g., mount, umount) are
always trapped by sysbox regardless of whether it operates in sys container
mode or not. This is a security feature to ensure initial container mount
immutability as well as emulation of newly mounted procfs/sysfs inside
the container.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>