Improve error reporting on bind-mount errors.
Prior to this change, when any error occurred during a bind-mount into a
container, sysbox-runc would retry the operation using the "mount" command
(i.e., instead of just the mount syscall). This was done because we've seen
issues bind-mounting when using shiftfs. However, it's a bit broad because we
only need to try the mount command when we get permission errors (EPERM) rather
than all errors.

Furthermore trying the mount command on all errors was causing sysbox-runc to
return an error back to the the higher-level container manager (e.g., Docker)
which was not always decoded properly by Docker.

For example, prior to this change, using Docker + Sysbox to mount a host file
into a container directory (an invalid operation) would fail with an error such
as:

"bind-mount of <host-file> to <container-dir> failed: exit status 32: unknown.

After this change the error message is more clear:

"bind-mount of <host-file> to <container-dir> failed: not a directory: unknown:
Are you trying to mount a directory onto a file (or vice-versa)? Check if the
specified host path exists and is the expected type."

That is, with this commit, sysbox-runc returns a proper error code to Docker,
which in turn decodes it and outputs a helpful message.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>