Add support for ID-mapping of rootfs.
This commit adds support for ID-mapping the container's rootfs,
This provides a better alternative to shiftfs (scheduled for
deprecation) and chowning (slow).

With this commit, sysbox-runc will use ID-mapping for the
container's rootfs when supported by the host kernel. In the
common case where the rootfs is on overlayfs, this requires
kernel 5.19 (sysbox-mgr is in charge of checking if the kernel
in fact supports ID-mapping on overlayfs).

If the kernel does not support ID-mapping on the container's
rootfs, Sysbox will resort to shiftfs (if supported by the kernel)
or chowing (last resort).

Notes:

* When ID-mapping the container's rootfs on overlayfs, the ID-mapping occurs on
  the lower-layers of the overlayfs, not in the upper layer. Thus, the if the
  container is stopped and restarted, it must re-use the same user-ns
  ID-mapping. In Sysbox-CE, this is currently guaranteed given that all
  containers use the same ID-mapping.

* The ID-mapping of the overlayfs lower layers occurs within the mount namespace
  of the container. Thus it's not visible at host level, and not even visible
  inside the container given that the lower-layers are outside the container's
  chroot jail. It's effect however is visible in the container given that the
  rootfs will show the correct uid:gid inside the container.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>