When ID-mapping rootfs, must chown the container's /
When using ID-mapping on the container's rootfs, the ID-mapping
applies to all files and dirs under the container rootfs, but
not the the container's rootfs itself (i.e., not to "/").
Therefore, operations such as "mkdir /somedir" fail with EPERM,
while operations such as "mkdir /root/somedir" work fine.

To fix this, sysbox-runc must chown the container's rootfs dir
(only the top dir, not it's subdirs) to match the container's
user-namespace user:group ID mappings.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>