Fix bug in uid shifting of implicit mounts.
There are two scenarios where we want to skip uid shifting of implicit
mounts done by sysbox-mgr:

1) When the rootfs is id-mapped (via ID-map-mounts or shiftfs) and
the kernel supports overlayfs on ID-mapped mounts (since overlayfs
will be mounted inside those implicit mounts).

2) When the rootfs is not id-mapped (e.g., docker userns-remap mode).
In this case we don't need to ID-map the implicit mounts since we
want them to have the same uid:gid ownership as the rootfs.

In this commit we fix a bug to ensure (2) works.

As an aside, here is the case where we want to uid-shift the implicit mounts:

* When the rootfs is on ID-mapped (e.g., shiftfs), but overlayfs on ID-mapping
  does not work. In this case, when the container starts or stops and we
  transfer contents from the implicit mount to the container's rootfs, we must
  chown since the rootfs will have uid:gid in the range 0->64K, while the
  implicit mount will have it in the range user-ns uid:gid (e.g.,
  100000->165535).

Bonus: update libsysbox linux capabilities list to accomodate new caps
       (CAP_PERFMON, CAP_BPF, CAP_CHECKPOINT_RESTORE).

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>