Fix rootfs ID-mapping bug exposed by docker -w.
When the rootfs is on overlayfs and Sysbox ID-maps it (kernel 5.19+),
the rootfs upper layer (which can't be ID-mapped) must be chowned.
Prior to this commit Sysbox was not doing this.

In the past we didn't catch this bug because the upper layer is usually empty,
so there was nothing to chown. But a command such as "docker
--runtime=sysbox-runc -w /dir <image>" will populate the upper layer with a
"/dir" directory. Without Sysbox changing it's ownership, that dir will show up
as "nobody:nogroup" inside the container.

Note: changes where also required in sysbox-mgr to account for the fact
that sysbox-runc will now chown the rootfs overlay upper dir, so that
sysbox-mgr can revert the chown properly during container stop or pause.
Refer to the sysbox-mgr repo for those changes.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>