Fix a bug in Sysbox's seccomp config.
The bug was causing Sysbox to not allow syscalls that should be
allowed inside system containers. The bug was exposed when we
updated the CRI-O version in the test container, since the new
CRI-O version creates the container's seccomp config in a
slightly more complex way than prior versions of CRI-O did.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>