Ensure sysbox-runc handles relative paths in overlayfs lowerdir option.
When sysbox-runc uses idmapped mounts on overlayfs, it does so by
unmounting overlayfs from the container's rootfs, then idmapping each
of the overlayfs lower layers, and then remounting the overlayfs into
the container's rootfs.

Turns out that for container images that have a very large number of lower
layers, the overlayfs "lowerdir" paths may be relative rather than absolute
(e.g., lowerdir=some/path:some/otherpath..."). This happens when the lowerdir
paths are so many that using absolute paths would exceed the number of
characters Linux allows on filesystem options.

This commit modifies the sysbox-runc code to deal with relative paths in
the lowerdir option of the container's rootfs overlayfs mount.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>